move the configuration to a GlobalConfigation object
