+++ src/main/resources/io/jenkins/plugins/artifact_manager_s3/S3BlobStore/config.properties
@@ -1,0 +1,1 @@
+configure_first=You have to configure your Amazon S3 setting in the section "Amazon S3 Bucket Access settings"
