+++ src/main/resources/io/jenkins/plugins/artifact_manager_s3/S3BlobStoreConfig/help-container.html
@@ -1,0 +1,1 @@
+<div>Name of the S3 Bucket, the S3 Bucket should existe and the AWS acount/profile/role used to access should have access to it</div>
