+++ src/main/resources/io/jenkins/plugins/artifact_manager_s3/S3BlobStoreConfig/help-deleteblods.html
@@ -1,0 +1,1 @@
+<div>this parameter is set by the property "-Dio.jenkins.plugins.artifact_manager_s3.S3BlobStoreConfig.deleteBlobs" enable delete blobs on a build is deleted</div>
