+++ src/main/resources/io/jenkins/plugins/artifact_manager_s3/S3BlobStoreConfig/help-prefix.html
@@ -1,0 +1,1 @@
+<div>This is the string that will be used as prefix for every file path, it must end with an "/" if it is a folder.</div>
