+++ src/main/resources/io/jenkins/plugins/artifact_manager_s3/S3BlobStoreConfig/help-region.html
@@ -1,0 +1,1 @@
+<div>Force the regios used to generate the URLs to access to the artifacts, if it is not set the plugin try to detect it.</div>
